Implement templating within containerTemplate.
Before, ApplyReplacements() would apply templating replacements to the containers within spec.Steps, but never to the single container at spec.ContainerTemplate. Note that this didn't work earlier because containerTemplate is merged with the container spec after templating is applied.

Isolate the container-specific templating in applyContainerReplacements()
and apply that logic to the containerTemplate field.

This addresses #811.