Refactor TaskParam and PipelineParam.
Make a new ParamSpec type and place it in its own file along with the
already existing Param type. ParamSpec replaces TaskParam and
PipelineParam.

TaskParam and PipelineParam are identical types, and with future
features including array/string parameters, consolidating
parameter-related code in param_types.go makes sense for better
code-reusability and logical structure.

Additionally, I think "ParamSpec" is a better name; TaskParam and
PipelineParam currently represent a name and description (with
potentially more information in the future such as whether the parameter
is an array or string), while the Param type represents the parameter
value itself. Previously, it was not as intuitive that Param contains
the actual parameter value, not TaskParam/PipelineParam.