Use moar builders in tests 👷‍♀️
In #1043 @EliZucker updated the package names of the v1alpha1 tests to
be `v1alpha1_test`. He did this b/c he wanted to use a function he was
adding to the builder test package for the work he is doing in #207, but
this has the nice side effect of making it so that we can use more
builders in these tests! (We couldn't before b/c the builders themselves
have to import and use `v1alpha1`). This commit updates many of these
tests to use builders.