Use our own completion for zsh
We are not doing the bash emulation since it's somewhat buggy (and slow) with aliases.
We are not doing the zsh output from cobra which while improved is still limited
for our cases
We are doing our own because we have much  more features (description) and much
flexible for our cases. It would need some manual update tho, but will work in a
further release to automate some of the heavy lifting.

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>