Expand parameters to support the Array type.
Instead of just strings, add support for users to supply parameters in
the form of an array. Also include appropriate type validation and
defaulting to support backwards-compatibility. Fixes #207.

The size of a user-supplied array is not specified anywhere in the
corresponding ParamSpec, which is useful for defining tasks that
require a dynamic number of strings used as distinct elements in
an array. For instance, one might use an array parameter in an
image-building task that feeds in a dynamic number of flags via the
args field for a particular step.

The implementation is defined such that an Array parameter can only be
used if the replacement string is completely isolated and within a field
that is an array of strings (currently eligible fields are 'command'
and 'args'). For instance, the webhook will prevent an array parameter
from being used in the 'image' field. Similarly, an array parameter used
in a composite string, like 'value=${array}', is invalid and will be
caught. The decision was made to completely prevent any use of array
parameters as strings because it clutters Tekton with unnecessary
functionality and is naturally extensible beyond the intended scope.
For instance, if the behavior was defined such that array parameters
used in the context of strings were automatically converted by
separating them with a space, then that relatively arbitrary
implementation could be questioned and eventually lead to needing to
specify a custom delimiter (like commas), and so on.

Another implementation detail to note is that the ArrayOrString type was
necessary, in combination with implementing json.Marshaller, to support
backwards-compatibility with strings. This is based off IntOrString from
kubernetes/apimachinery, and allows the value-type to be immediately
parsed and decided through the webhook.

Lastly, a possibly unintuitive design decision is that if an EMPTY array
parameter properly replaces an array element, then that string in the
original array will be completely removed.