Enhance the format of the Cloud Event sent by Tekton
Move TaskRun to a TaskRun field in the payload to prepare for
extra data we may want to add in the future without breaking
existing consumers.

Add a version number to the message types so that if at some
point we want to start sending a new format of message we can
match it to a new version.

Signed-off-by: Andrea Frittoli <andrea.frittoli@uk.ibm.com>