Migrate/Upgrade previous version with a default timeout…
In case of a controller upgrade, we may still store previous versions
of resource that may not have been the default set. This migrates
those "in memory" to make sure we do not panic in those cases.

This also gives a great example of how to use a mutating webhook to
migrate some resources from one version to another.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>