Change the behavior of outputs that are also used as inputs.
This change makes the handling of Resources within a Task consistent, regardless of
whether the same Resource is used as both an input and an output. Previously these
were special cased, which made it hard to write Tasks consistently.

This commit also makes a few minor changes to the way our bash output gets logged.
I discovered this was missing during debugging, and made it consistent with the gsutil
wrapper.

This is a followup to https://github.com/tektoncd/pipeline/pull/1119 and should be submitted
once the next release is cut.