Reduce the number of imageDigestExporter step to one per Task
Currently, if a `Task` uses one or more image output we add a
`imageDigestExporter` for each step defined in the `Task`. This is not
required as only one step/container can do the work and handle
multiple images.

This change reduce to one the number of `imageDigestExporter` to one
step, after the last user-defined step.

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>