Add Condition type to webhook and `SchemeBuilder`
To enable webhook validation, Condition has to implement the
`Defaultable` interface even though at the moment there are no
default values.
Also, found that we need to register new types to the SchemeBuilder else
if we cannot create Conditions using the generated client (e.g. for e2e
 tests) since the Kind is not properly set.

Both of these changes should have been a part of the APIs PR for
Conditions. I will have an upcoming PR that will update docs for creating new
types with this information.

Signed-off-by: Dibyo Mukherjee <dibyo@google.com>