Refactor input resource volume handling to remove a type switch statement.
This change refactors the way input resources attach volumes to their steps.
Previously, there was a special-cased switch statement for Storage type resources.
This made the logic for resources hard to follow, and spread across the codebase. This
change refactors that by moving it up into the general Resource interface.

This is another follow-on to https://github.com/tektoncd/pipeline/pull/1135