Automatically create named directories for each output resource in a TaskRun.
TaskRuns are expected to place files related to an Output in a directory named
for each Output. Right now, we expect Task authors to create this directory before
placing files there. This is onerous and error-prone, so we should automatically create
them.

There are a few approaches to doing this:
- a mkdir init container
- a volume
- a mkdir pod container

I chose to use the final one, a mkdir pod container for consistency with Input resources.
This is prepended to the Task steps.