Fix staleness issue in entrypoint resolver.
There was a bug in the entrypoint resolution logic that allowed for stale values of the
entrypoint to be returned. If an image is supplied by tag, and that tag changes to point
to an image with a new entrypoint, we returned the old one.

This change uses the image digest itself as our cache key, instead of the image name.