Update fan in / fan out test to use path variable replacement 👣
Using the path replacement makes dealing with paths much easier - you
don't need to know about `/workspace` or about when the path has been
overridden or any of that.

I'm also doing this to set the stage for #850 - so that when we change
from `${}` syntax to `$()` we can use this test to test for both types
(to make sure the change is backward compatible).