Add support for $() variable substitution in addition to ${}
In #850 we decided that to make our syntax more similar to k8s style
syntax, we will use $() for variable substitution instead of ${}. In a
later iteration we may also want to make it so that anything that can be
acesssed as a variable is also available as an env var to the running
container, but that is TBD.

In the context of this discussion we also considered removing variable
substitution entirely, that conversation is continuing.

We have also tried to start calling this functionality "variable
substitution" instead of "templating" to make it clear that we do not
intend to support more complicated templating functionality.

Co-authored-by: Eli Zucker <thezuck@google.com>