Document API change policy 👩 ⚖️
This commit adds docs about the policy the current governing board has been
discussing around how to handle API changes and make sure they are
reviewed by the ppl that should review them.

We can adjust this policy as needed, e.g. if we start running into
situations where not enough owners can review within a reasonable amount
of time.

Also added links to the doc where we are working on our plan to get to
Beta and updated our docs about this.

Fixes #906