Fix pipelinerun status update when tasks are skipped
Before conditionals, a pipelinerun was marked successful only
if all tasks were successfully run. Now, a pipelinerun is successful
if all tasks were either successfully run or skipped due to a condition
check failure.

Currently while updating status, we only see if a task's own condition
check failed. This is not sufficient if that task (B) is dependent on another
task (A) using `from` or `runAfter`. If A is skipped due to a condition failure,
task B is not scheduled to be execute, and there are no other tasks left to run.
However, the pipeline status is never updated from `Running` to `Successful`.

To fix this, this commit changes the `GetPipelineConditionStatus` function to
recursively look at a task's parent tasks to see if they are skipped. If any single
parent task is skipped, we skip the task. If all tasks in a pipeline are either
successful or skipped, we update the pipeline run status to successful.

Signed-off-by: Dibyo Mukherjee <dibyo@google.com>