Change the interface of the PullRequest resource to be more granular.
Currently the PullRequest resource produces one large json object representing the
entire PR. This works, but makes it harder to write simple programs to manipulate this
object.

This commit changes this interface to produce a set of files, each representing a part
of the initial pull request. This makes it easier for programs to do simple tasks like add
labels or comments.