Propagate labels and annotation early on for PipelineRun
If a PipelineRun fails on validation, with `ReasonInvalidGraph`,
`ReasonInvalidBindings` or `ReasonParameterTypeMismatch`, the
PipelineRun do not inherit the referenced Pipeline labels and
annotation.

This poses a problem for tools that uses those labels, especially
`tekton.dev/pipeline=` to filter the PipelineRun for a given
Pipeline.

This commits fixes that by propagating those labels and annotation as
early as it can (aka after making sure the PipelineRef is valid)

Signed-off-by: Vincent Demeester <vdemeest@redhat.com>