Add links to versioned docs from main README 🔗
For a long time now we've wanted to have versioned docs - the docs you
get when you look at the repo only work for HEAD, which can be different
from our releases.

For example, TWICE today I've seen users try to run our examples with
the new `$()` syntax with `v0.5.2` deployed :( @vdemeester pointed out
we should get versioned docs asap so until we have some cool docs
generated (with versions!) at tekton.dev, I think this might be an okay
stopgap?

This commit adds a table which links to the markdown files associated
with each version. I'm also going to update the releases in GitHub
themselves to contain a link to these docs and examples.