WIP: Initial implementation of step scripts
This adds a `script` field to the `Step` type which, when specified,
results in a temporary generated executable script file being generated
and invoked containing the specified script items.

The result is an easier-to-use scripting option for users who just want
to invoke simple scripts. Generated scripts should be an implpementation
detail and will not persist between steps.