Implement step scripts
This adds a `Script` field to the `Step` type which, when specified,
results in a temporary generated executable script file being invoked
containing the specified script contents.

The result is an easy-to-use scripting option for users who just want
to invoke simple scripts inside containers. Details of generated scripts
should be considered an implementation detail, and should not be relied
upon by users.