Update v0.6.0 docs to include fixed tutorial 🤒
In #1199 @moficodes reported that the tutorial wasn't working which was
b/c in #1172 when I updated the braces, I only half updated the tutorial
example. I wanted to fix this for the v0.6.0 docs as well, but since the
released binary and yaml files don't need an update, it didn't make
sense to me to create v0.6.1 for this (and leave the v0.6.0 docs
broken!) so I created a 0.6 branch and cherry picked the docs fix into
it only, and am now updating the README to point at that.