Add a new method to the PipelineResource interface for customization of the Output directory
We previously initialized each Output Resource with an empty directory. This changes the
behavior to initialize this with a custom set of files per resource. The behavior now is
the same as before, with each resource embedding a helper to create empty directories.

Over time we can customize this where necessary.