Explicitly substitute resource conditional variables.
Previously, we'd change the variable names to what the
taskrun format and then rely on the taskrun's variable
substitution logic. This approach is hacky and now we
explicitly replace the resource variable strings with the
right values without relying on the TaskRun logic.

In the future, we should consider doing the same for parameter
variables as well.

Signed-off-by: Dibyo Mukherjee <dibyo@google.com>