Add version to release tarball
We don't currently have a VERSION in the release tarball. Which breaks `tkn
version` when building from sources or with package managers (i.e: brew, deb,
rpm etc.. )

On gorelease we grab the curren tag from git which should be defined in the
git pipelineresource and output it to the VERSION file.

Make would then see this file and add the proper flags if he sees it. (the
LDFLAGS can be overrided if needed).

Closes #231

Signed-off-by: Chmouel Boudjnah <chmouel@redhat.com>