Accept any sidecar termination Reason as success
Fixes #1253

When a sidecar is forcefully exited by the taskrun controller it may
error out. This happens because the sidecar is terminated by having its
image field rewritten to be tekton's nop container. The nop container is
bare bones and probably doesnt contain the entrypoint command configured
for the sidecar. When this occurs the container will exit with an error
of some kind. On GKE this error's Reason is "ContainerCannotRun" but on
MiniShift/OpenShift the error Reason is "Error".

The sidecar e2e test previously required a termination Reason of
"ContainerCannotRun" or "Completed" but this appears to be too rigid.
Instead the test now considers a sidecar that has been terminated for
any reason to be a success.