Fix interactive prompt with --last
This will fix the issue of interactive prompt getting
started with --last flag in scenario of not using
resource or param flag. It should take the value from
last pipelinerun in case of no use of resource and
param flag. In case of using the flag, it will override
the value of last run but should never start
interactive prompt.

Change the resource selection question in interactive
prompt for better user experience.

Adds tests for the given scenario

Fixes https://github.com/tektoncd/cli/issues/297