Show descriptive failure message
This will fix the issue of failure message shown in
pipelinerun describe was less descriptive

We were showing the failure message from condition
of pipelinerun status which was like this task failed

This was not actually describing the failure reason

Now we will show the failure message from the
condition of the failed taskrun status.
This will show the actual reason of failure

Format will be prMsg(trMsg)

Added tests for that.

Fixes https://github.com/tektoncd/cli/issues/64