Remove logging from timeout handler ✏️
Logging in the timeout handler was added as part of #731 cuz it helped
us debug when the timeout handler didn't work as expected. Unfortunately
it looks like the logger we're using can't be used in multiple go
routines (https://github.com/uber-go/zap/issues/99 may be related).
Removing this logging to fix #1124, hopefully can find a safe way to add
logging back in #1307.