Include vendored source in release-built images
This adds logic to the nightly release Task that targz's up everything
in vendor/ and includes it in ko-built container images. Some of our
dependencies' licenses require their source to be included in
distributed artifacts (like container images).

Once we've determined this works fine for nightly releases, I'll copy
this to publish.yaml so it's also done for official releases.