Use kubectl create instead of apply
When running YAML E2E tests, use kubectl create instead of apply,
so that example can use generateName, which is a best practice
for run objects (TaskRun/PipelineRun).

Using "create" also exposed name conflicts for several resources
all named the same. Luckily the content of those secret was the
same, so we had no race condition in the past, but it's best
to avoid side effects between tests as much as possible and
keep resource names uniques across examples.

With this change we start running taskrun tests again, and it
turns out the gitlab test is broken because of a broken gitlab URL.
Introduce a subfolder for examples that are valid but that we
don't want to run in CI because they need some access to a
resource which is not available.

The Gitlab test has been removed before this was merged, but
keeping the no-ci folder anyways as it may be useful for examples
that we have no easy way to run in CI.

Fixes #1397