Helpful error message when multiple volumes share name
A Task is not allowed to have multiple volumes that share a name. Prior
to this commit the error message printed when a name collision occurred
was "expected exactly one, got both: name" which doesn't describe the
problem clearly.

After this commit the error message is updated to 'multiple volumes with
same name "foo"'. A test case has been added to exercise the new message.