Change the path used by the image digest exporter to the standard output path structure.
Previously we were writing output digests to /builder/home/image-output-paths. This commit
unifies that system and expects tasks to output these files to /workspace/output/<resource name>.

This does introduce a change in behavior, but support is retained for the old location. A warning
is logged if the binary finds an index.json in the old location.