Use a local registry in build-push-kaniko
This is a first step in making our E2E tests more portable and less
dependent on resources outside of the k8s cluster they run into.

This changes the build-push-kaniko taskrun example.
There is a lot of infra setup required to make sure that the service
account used by test is able to push to a registry. The test script
replaces the image URL in the resource with the content of the
KO_DOCKER_REPO environment variable. If that is a gcr.io url then
the service account must be setup for access to it.

The updated version runs a local registry as a sidecar in the pod.
The task now can push to localhost:5000. The port is not exposed
outside of the pod so there is no risk of interference with other
tests.

Partially fixes #1372