Migrate PullRequestResource to go-scm.
This changes the underlying implementation of the PullRequestResource
to jenkins-x/go-scm in order to take advantage of the repo drivers to
support other SCM providers.

In the process, this also allowed for refactoring to have a cleaner
distinction between interaction with SCM APIs and converting the data to
an on disk format.

BREAKING CHANGES:
* go-scm does not yet support updating comments, so that behavior is
removed for the time being since:
  1) it's unclear if anyone is actively
  using it.
  2) this behavior can be roughly replicated with a delete +
  create.
  3) the benefits of having the other SCM providers outweighs this
  feature.

Fixes #1066.