Add a tekton v0.7.0 version of the nightly release
Create a new version of the nightly-release that is identical to
the full release, except for the following:
- no release pre-check is executed
- a new ko image is built (using a v0.7.0 compatible task)
- the default image registry is gcr.io/tekton-nightly

This version can be executed via triggers on the dogfooding
cluster. This commit leaves the existing nightly release untouched
so that it can be executed via prow, until we are ready to
finalize the switch to the dogfooding cluster for nightly.