Enhance the cluster resource type a bit.
I was trying it out with a GKE cluster, and noticed a few issues:
- we require a username for everything, even if we're using CA auth.
- we weren't checking token auth.

This should really be switched to a declarative system where resources can
specify optional params, but that can come as part of the larger refactor.
This makes things a bit nicer tot use in the meantime.