Make some docs a smidge more readable 👓
Moved the links to the list of PipelineResources that we support further
up in the doc so ppl casually glancing at it know its there, and ppl like
me who refer to it all the time can get to it right away.

Added some syntax highlighting around parameters in the install doc so
they stand out from the surrounding text more.